# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Diary Service API
  description: plip-diary REST API
  version: v1.0.0
servers:
- url: http://localhost
  description: Generated server url
tags:
- name: Theme
  description: 다이어리 테마 API
- name: Home
  description: 홈 피드 API
- name: Calendar
  description: 캘린더 API
- name: Diary Video
  description: 다이어리 영상 API
- name: Timeline
  description: 다이어리 타임라인 API
paths:
  /api/v1/diaries/videos/{diaryVideoId}/topic-transfer:
    post:
      tags:
      - Diary Video
      summary: 다이어리 영상 → 토픽 복사·이동
      description: "topic-service에 영상 등록이 끝난 뒤(MOVE), 다이어리 타임라인에서 제거. `diary.video.unlinked`\
        \ 발행. COPY는 topic-service API 사용."
      operationId: transferDiaryVideoToTopic
      parameters:
      - name: diaryVideoId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferDiaryVideoToTopicRequest"
        required: true
      responses:
        "200":
          description: OK
      security:
      - bearerAuth: []
  /api/v1/diaries/themes:
    get:
      tags:
      - Theme
      summary: 테마 목록 조회
      description: 사용자의 활성 테마 목록 반환
      operationId: listThemes
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ThemeListResponse"
      security:
      - bearerAuth: []
    post:
      tags:
      - Theme
      summary: 테마 생성
      description: 활성 테마는 사용자당 최대 5개까지 생성 가능
      operationId: createTheme
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateThemeRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ThemeResponse"
      security:
      - bearerAuth: []
  /api/v1/diaries/themes/{id}:
    get:
      tags:
      - Theme
      summary: 테마 메타 단건 조회
      description: "테마 이름·생성일 등 **메타 정보만** 반환. 테마에 포함된 영상 목록은 `GET /api/v1/diaries/themes/{id}/timeline`\
        \ 사용."
      operationId: getTheme
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ThemeResponse"
      security:
      - bearerAuth: []
    delete:
      tags:
      - Theme
      summary: 테마 삭제
      description: 테마와 연관 영상을 Soft Delete 처리
      operationId: deleteTheme
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
      security:
      - bearerAuth: []
    patch:
      tags:
      - Theme
      summary: 테마 이름 수정
      operationId: updateTheme
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateThemeRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ThemeResponse"
      security:
      - bearerAuth: []
  /api/v1/diaries/themes/{id}/timeline:
    get:
      tags:
      - Timeline
      summary: 테마별 다이어리 조회
      description: "특정 테마의 영상을 날짜별 섹션으로 반환. `diary_videos` + Read Model(캡션·썸네일) enrichment.\
        \ 테마 메타는 `GET /api/v1/diaries/themes/{id}` 별도 호출."
      operationId: getThemeTimeline
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ThemeTimelineResponse"
      security:
      - bearerAuth: []
  /api/v1/diaries/home:
    get:
      tags:
      - Home
      summary: 홈 피드 조회
      description: "KST 오늘(0건이어도 포함) + 영상이 있는 최신 2일, 일자별 최대 4건. `diary_videos` + Read\
        \ Model(캡션·썸네일) enrichment. 테마 전체 목록은 `GET /api/v1/diaries/themes` 별도 호출.\
        \ 특정일 전체는 `GET /api/v1/diaries/dates/{date}`."
      operationId: getHomeFeed
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/HomeFeedResponse"
      security:
      - bearerAuth: []
  /api/v1/diaries/dates/{date}:
    get:
      tags:
      - Timeline
      summary: 날짜별 다이어리 조회
      description: 특정일(KST) 영상을 테마별 섹션으로 반환. `diary_videos` + Read Model(캡션·썸네일) enrichment.
        테마 전체 목록은 `GET /api/v1/diaries/themes` 별도 호출.
      operationId: getDateTimeline
      parameters:
      - name: date
        in: path
        required: true
        schema:
          type: string
          format: date
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DateTimelineResponse"
      security:
      - bearerAuth: []
  /api/v1/diaries/calendar:
    get:
      tags:
      - Calendar
      summary: 월별 작성일 조회
      description: 캘린더 dot 표시용 작성일 목록. 활성 테마·영상만 포함. 영상 메타·썸네일 미포함.
      operationId: getCalendar
      parameters:
      - name: year
        in: query
        required: true
        schema:
          type: integer
          format: int32
      - name: month
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CalendarResponse"
      security:
      - bearerAuth: []
  /api/v1/diaries/videos/{diaryVideoId}:
    delete:
      tags:
      - Diary Video
      summary: 다이어리 영상 바인딩 해제
      description: 타임라인에서 영상 항목 제거. `diary_videos` Soft Delete 후 `diary.video.unlinked`
        Kafka 발행.
      operationId: unbindDiaryVideo
      parameters:
      - name: diaryVideoId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
      security:
      - bearerAuth: []
components:
  schemas:
    TransferDiaryVideoToTopicRequest:
      type: object
      description: 다이어리 영상 → 토픽 이동 요청 (diary 담당 구간)
      properties:
        mode:
          type: string
          description: "MOVE: topic-service 바인딩 완료 후 다이어리 타임라인에서 제거. COPY는 topic-service\
            \ API 사용"
          enum:
          - COPY
          - MOVE
      required:
      - mode
    CreateThemeRequest:
      type: object
      description: 테마 생성 요청
      properties:
        name:
          type: string
          description: 테마 이름
          example: 여행
          maxLength: 50
          minLength: 0
      required:
      - name
    ThemeResponse:
      type: object
      description: 테마 응답
      properties:
        id:
          type: integer
          format: int64
          description: 테마 ID
          example: 1
        themeUuid:
          type: string
          format: uuid
          description: 테마 UUID (cross-service 식별자)
          example: 01912345-6789-7abc-def0-123456789abc
        name:
          type: string
          description: 테마 이름
          example: 일상
        createdAt:
          type: string
          format: date-time
          description: 생성 시각
        updatedAt:
          type: string
          format: date-time
          description: 수정 시각
    UpdateThemeRequest:
      type: object
      description: 테마 수정 요청
      properties:
        name:
          type: string
          description: 테마 이름
          example: 여행 기록
          maxLength: 50
          minLength: 0
      required:
      - name
    ThemeListResponse:
      type: object
      description: 테마 목록 응답
      properties:
        themes:
          type: array
          description: 활성 테마 목록
          items:
            $ref: "#/components/schemas/ThemeResponse"
    ThemeTimelineResponse:
      type: object
      description: 테마별 다이어리 타임라인 응답
      properties:
        sections:
          type: array
          description: "날짜별 영상 섹션 (영상이 있는 날짜만, 최신 날짜 우선)"
          items:
            $ref: "#/components/schemas/ThemeTimelineSectionResponse"
    ThemeTimelineSectionResponse:
      type: object
      description: 테마별 타임라인 날짜 섹션
      properties:
        date:
          type: string
          format: date
          description: "작성일 (KST, ISO-8601 date)"
          example: 2026-08-01
        videos:
          type: array
          description: 해당 날짜·테마의 영상 목록
          items:
            $ref: "#/components/schemas/ThemeTimelineVideoResponse"
    ThemeTimelineVideoResponse:
      type: object
      description: 테마별 타임라인 영상 항목
      properties:
        id:
          type: integer
          format: int64
          description: 다이어리 영상 ID
          example: 10
        videoUuid:
          type: string
          format: uuid
          description: "영상 UUID (video-service 발급, diary 참조)"
          example: 01912345-6789-7abc-def0-123456789abd
        caption:
          type: string
          description: 캡션 (Read Model enrichment)
        thumbnailUrl:
          type: string
          description: 썸네일 URL (Read Model enrichment)
        createdAt:
          type: string
          format: date-time
          description: 바인딩 시각
    HomeFeedResponse:
      type: object
      description: 홈 피드 응답
      properties:
        sections:
          type: array
          description: "날짜별 영상 섹션 (오늘 KST 필수 포함, 최대 3일, 최신 날짜 우선)"
          items:
            $ref: "#/components/schemas/HomeFeedSectionResponse"
    HomeFeedSectionResponse:
      type: object
      description: 홈 피드 날짜 섹션
      properties:
        date:
          type: string
          format: date
          description: "작성일 (KST, ISO-8601 date)"
          example: 2026-08-15
        videos:
          type: array
          description: "해당 날짜의 영상 목록 (최대 4건, 최신순)"
          items:
            $ref: "#/components/schemas/HomeFeedVideoResponse"
    HomeFeedVideoResponse:
      type: object
      description: 홈 피드 영상 항목
      properties:
        id:
          type: integer
          format: int64
          description: 다이어리 영상 ID
          example: 10
        themeId:
          type: integer
          format: int64
          description: 테마 ID
          example: 1
        themeName:
          type: string
          description: 테마 이름
          example: 일상
        videoUuid:
          type: string
          format: uuid
          description: "영상 UUID (video-service 발급, diary 참조)"
          example: 01912345-6789-7abc-def0-123456789abd
        caption:
          type: string
          description: 캡션 (Read Model enrichment)
        thumbnailUrl:
          type: string
          description: 썸네일 URL (Read Model enrichment)
        createdAt:
          type: string
          format: date-time
          description: 바인딩 시각
    DateTimelineResponse:
      type: object
      description: 날짜별 다이어리 타임라인 응답
      properties:
        date:
          type: string
          format: date
          description: "조회 날짜 (KST, ISO-8601 date)"
          example: 2026-08-01
        sections:
          type: array
          description: 테마별 영상 섹션 (해당 날짜에 영상이 있는 테마만)
          items:
            $ref: "#/components/schemas/DateTimelineSectionResponse"
    DateTimelineSectionResponse:
      type: object
      description: 날짜별 타임라인 테마 섹션
      properties:
        themeId:
          type: integer
          format: int64
          description: 테마 ID
          example: 1
        themeName:
          type: string
          description: 테마 이름
          example: 일상
        videos:
          type: array
          description: 해당 테마·날짜의 영상 목록
          items:
            $ref: "#/components/schemas/DateTimelineVideoResponse"
    DateTimelineVideoResponse:
      type: object
      description: 날짜별 타임라인 영상 항목
      properties:
        id:
          type: integer
          format: int64
          description: 다이어리 영상 ID
          example: 10
        videoUuid:
          type: string
          format: uuid
          description: "영상 UUID (video-service 발급, diary 참조)"
          example: 01912345-6789-7abc-def0-123456789abd
        caption:
          type: string
          description: 캡션 (Read Model enrichment)
        thumbnailUrl:
          type: string
          description: 썸네일 URL (Read Model enrichment)
        createdAt:
          type: string
          format: date-time
          description: 바인딩 시각
    CalendarResponse:
      type: object
      description: 월별 캘린더 인디케이터 응답
      properties:
        year:
          type: integer
          format: int32
          description: 조회 연도
          example: 2026
        month:
          type: integer
          format: int32
          description: 조회 월
          example: 8
        dates:
          type: array
          description: "영상 작성일 목록 (KST, ISO-8601 date)"
          items:
            type: string
            format: date
  securitySchemes:
    bearerAuth:
      type: http
      description: "Gateway JWT 인증. 클라이언트는 Bearer 토큰을 전달하고, Gateway가 검증 후 downstream에\
        \ X-User-Uuid를 주입합니다."
      scheme: bearer
      bearerFormat: JWT